Refactor no nome das classes Rest Refactor de atributos da entity Movie
