+++ src/main/java/br/com/iworks/movie/model/entity/Movie.java
@@ -12,12 +12,6 @@ import org.springframework.data.mongodb.core.index.CompoundIndex;
 import org.springframework.data.mongodb.core.index.CompoundIndexes;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-
-import br.com.iworks.movie.config.util.JsonDateTimeSerializer;
 import br.com.iworks.movie.model.GenreEnum;
 import br.com.iworks.movie.model.TypeEnum;
 import lombok.Data;
@@ -32,13 +26,11 @@ import lombok.NoArgsConstructor;
         @CompoundIndex(name = "code_movie_idx", def = "{'code': 1}", unique = true),
         @CompoundIndex(name = "title_year_movie_idx", def = "{'title': 1, 'year': 1}", unique = true, sparse = true)
 })
-@JsonIgnoreProperties(ignoreUnknown = true)
 public class Movie {
 
     public static final String COLLECTION_NAME = "movies";
 
     @Id
-    @JsonIgnore
     private String id;
 
     private Long code;
@@ -46,7 +38,6 @@ public class Movie {
     @NotNull(message = "{validation.notnull}")
     private String title;
     private String originalTitle;
-
     private String duration;
 
     @NotNull(message = "{validation.notnull}")
@@ -55,34 +46,17 @@ public class Movie {
     private Date registrationDate;
     private Date releasedDate;
 
-    @Max(value = 9999, message = "{validation.size}")
     @Min(value = 1800, message = "{validation.size}")
+    @Max(value = 3000, message = "{validation.size}")
     private Integer year;
 
     @NotNull(message = "{validation.notnull}")
     private String plot;
     private String director;
+
+    @Min(value = 0, message = "{validation.size}")
+    @Max(value = 5, message = "{validation.size}")
     private int rating;
     private String imdbRating;
     private String imdbID;
-
-    @JsonSerialize(using = JsonDateTimeSerializer.class)
-    public Date getReleasedDate() {
-        return releasedDate;
-    }
-
-    @JsonFormat(shape = JsonFormat.Shape.OBJECT)
-    public void setType(TypeEnum type) {
-        this.type = type;
-    }
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING)
-    public TypeEnum getType() {
-        return type;
-    }
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING)
-    public List<GenreEnum> getGenres() {
-        return genres;
-    }
 }
