+++ src/main/java/br/com/iworks/movie/service/impl/MovieServiceImpl.java
@@ -1,13 +1,12 @@
 package br.com.iworks.movie.service.impl;
 
-import br.com.iworks.movie.exceptions.MovieException;
-import br.com.iworks.movie.model.entity.Movie;
-import br.com.iworks.movie.model.entity.QMovie;
-import br.com.iworks.movie.repository.MovieRepository;
-import br.com.iworks.movie.service.CounterService;
-import br.com.iworks.movie.service.MovieService;
-import com.querydsl.core.types.dsl.BooleanExpression;
-import lombok.extern.slf4j.Slf4j;
+import java.util.Date;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import javax.validation.ConstraintViolation;
+import javax.validation.Validator;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.context.i18n.LocaleContextHolder;
@@ -16,11 +15,15 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validator;
-import java.util.Calendar;
-import java.util.Set;
-import java.util.stream.Collectors;
+import com.querydsl.core.types.dsl.BooleanExpression;
+
+import br.com.iworks.movie.exceptions.MovieException;
+import br.com.iworks.movie.model.entity.Movie;
+import br.com.iworks.movie.model.entity.QMovie;
+import br.com.iworks.movie.repository.MovieRepository;
+import br.com.iworks.movie.service.CounterService;
+import br.com.iworks.movie.service.MovieService;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @Slf4j
@@ -44,7 +47,7 @@ public class MovieServiceImpl implements MovieService {
             this.validateMovie(movie);
 
             movie.setCode(counterService.getNextSequence(Movie.COLLECTION_NAME));
-            movie.setRegistrationDate(Calendar.getInstance().getTime());
+            movie.setRegistrationDate(new Date());
             movie = repo.save(movie);
         } catch (Exception e) {
             throw movieException(movie, e);
