+++ src/main/java/br/com/iworks/movie/ws/v1/MovieRestController.java
@@ -35,7 +35,7 @@ import io.swagger.annotations.ApiResponses;
 @RestController
 @RequestMapping("/movies")
 @Api("/movies")
-public class MovieRest {
+public class MovieRestController {
 
     @Autowired
     private MovieService service;
