+++ src/main/java/br/com/iworks/movie/ws/v1/TypeRestController.java
@@ -1,7 +1,7 @@
 package br.com.iworks.movie.ws.v1;
 
 import br.com.iworks.movie.model.GenreEnum;
-import br.com.iworks.movie.model.entity.Movie;
+import br.com.iworks.movie.model.TypeEnum;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
@@ -10,19 +10,19 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/genres")
-@Api("/genres")
-public class GenreRest {
+@RequestMapping("/types")
+@Api("/types")
+public class TypeRestController {
 
-    @ApiOperation(value = "Get the list of genres")
+    @ApiOperation(value = "Get the list of types")
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "Success", response = GenreEnum.class),
+            @ApiResponse(code = 200, message = "Success", response = TypeEnum.class),
             @ApiResponse(code = 404, message = "Not Found"),
             @ApiResponse(code = 400, message = "Bad Request")})
     @ResponseBody
     @ResponseStatus(HttpStatus.OK)
     @RequestMapping(method = RequestMethod.GET)
-    public GenreEnum[] list() {
-        return GenreEnum.values();
+    public TypeEnum[] list() {
+        return TypeEnum.values();
     }
 }
